refactor : divide user item component into userProfile and userItem
